// server/src/user/dto/user-statistics.dto.ts
export class UserStatisticsDto {
    id: number;
    username: string;
    fullname: string;
    dob: string;
    email: string;
    phone: string;
    score: number;
    avatar: string;
    googleAccountId: string;
    learntWordsCount: number;
    completedTopicsCount: number;
    inProcessTopicsCount: number;
  }
  